CoolingJacket supports outer wall, area factor
An outer liner material is needed for some stress analysis, so it can now be included as a CoolingJacket **kwarg.
Ribs in between channels are now considered when determining the coolant flow area, through the blockage_ratio **kwarg, which defaults to 0. This has no effect on heat transfer (other than changing coolant velocity).